refactor: middleware to .ts, add matcher so middleware doesn't run when not needed, remove unused stuff in join-form and slack.tsx
